Expect Dispatcher replies to be tuples, not lists
As mentioned in issue 176, according to the documentation one of the ways a
callback function may pass a reply back to a client is by returning a message
encapsulated thusly: `tuple(<Address>, <ArgValue>)`.

Whilst the code for the UDP servers does conform to this, the code to TCP
servers was expecting `list(<Address>, <ArgValue>)` in all cases.

This PR resolves that, updating the tests as well.